% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Open a Browser}
\usage{
browse(url)
}
\arguments{
\item{url}{[\code{character(1)}]\cr
URL to open.}
}
\description{
Opens a browser to browse the provided \code{url}.
The default browser can be set via options:
\describe{
  \item{\code{gtfo.browse.bin}:}{Binary to call.}
  \item{\code{gtfo.browse.args}:}{Arguments for \code{bin}. Use \dQuote{\%s} as placeholder for the url.}
}

If no option is set, \command{xdg-open} will be called on Linux, \command{open} on Mac OS and \command{start} on Windows.
These command open the associated default application.
}
\examples{
\dontrun{
# Open the R project web site
browse("https://www.r-project.org")
}
}
\references{
\url{https://github.com/freedesktop/xdg-utils}
}
\seealso{
\code{\link{term}} to start a terminal and \code{\link{fm}} to start a file manager.
}
