% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{gtfo-package}
\alias{gtfo}
\alias{gtfo-package}
\title{gtfo: Go to terminal or file manager}
\description{
Leave R by opening a terminal or file manager at a specified location.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mllg/gtfo}
  \item Report bugs at \url{https://github.com/mllg/gtfo/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com}

}
\keyword{internal}
