% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term}
\alias{term}
\title{Open a Terminal}
\usage{
term(path = getwd())
}
\arguments{
\item{path}{[\code{character(1)}]\cr
Path. Defaults to the current working directory as reported by \code{\link[base]{getwd}}.}
}
\description{
Opens a terminal emulator in directory \code{path}.
First tries to create a \command{tmux} split, then to open a new terminal.
The default terminal can be set via options:
\describe{
  \item{\code{gtfo.term.bin}:}{Binary to call.}
  \item{\code{gtfo.term.args}:}{Arguments for \code{bin}. Use \dQuote{\%s} as placeholder for the path.}
}

If no option is set, but the current R session runs inside \command{tmux}, a tmux split will be opened.
Otherwise, a new terminal will be started using the following presets:
\describe{
 \item{Mac OS:}{If called from \command{iTerm}, start \command{iTerm}. Else start \command{Terminal}.}
 \item{Linux:}{Try the follwing terminals and open the first one found: \command{kitty}, \command{alacritty}, \command{urxvt}, \command{gnome-terminal}, \command{konsole}, \command{xfce4-terminal}, \command{lxterminal}.}
 \item{Windows:}{Start powershell.}
}
}
\examples{
\dontrun{
# Terminal in R's working directory
term()

# Terminal in R's tempdir
term(tempdir())
}
}
\seealso{
\code{\link{fm}} to start a file manager and \code{\link{browse}} to start a browser.
}
