% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm.R
\name{fm}
\alias{fm}
\title{Open a File Manager}
\usage{
fm(path = getwd())
}
\arguments{
\item{path}{[\code{character(1)}]\cr
Path. Defaults to the current working directory as reported by \code{\link[base]{getwd}}.}
}
\description{
Opens a file manager to browse the provided \code{path}.
The default file manager can be set via options:
\describe{
  \item{\code{gtfo.fm.bin}:}{Binary to call.}
  \item{\code{gtfo.fm.args}:}{Arguments for \code{bin}. Use \dQuote{\%s} as placeholder for the path.}
}

If no option is set, \command{xdg-open} will be called on Linux, \command{open} on Mac OS and \command{start} on Windows.
These commands open the associated default application.
}
\examples{
\dontrun{
# Brose R's working directory
fm()

# Inspect R's tempdir
fm(tempdir())
}
}
\references{
\url{https://github.com/freedesktop/xdg-utils}
}
\seealso{
\code{\link{term}} to start a terminal and \code{\link{browse}} to start a browser.
}
