% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term}
\alias{term}
\title{Open a Terminal}
\usage{
term(path = getwd())
}
\arguments{
\item{path}{[\code{character(1)}]\cr
Path. Defaults to the current working directory as reported by \code{\link[base]{getwd}}.}
}
\description{
Opens a terminal emulator in directory \code{path}.
First tries to create a \command{tmux} split, then to open a new terminal.
The default terminal can be set via options:
\describe{
  \item{\code{gtfo.term.bin}:}{Binary to call.}
  \item{\code{gtfo.term.args}:}{Arguments for \code{bin}. Use \dQuote{\%s} as placeholder for the path.}
}

If no option is set, the terminal defaults to \command{iTerm} on Mac OS and \command{cmd.exe} on Windows.
On Linux, the first of the following terminal emulators found will be used:
\command{gnome-terminal}, \command{konsole}, \command{xfce4-terminal}, \command{kitty}, \command{alacritty},
\command{urxvt}, \command{lxterminal}.
}
\examples{
\dontrun{
# Terminal in R's working directory
term()

# Terminal in R's tempdir
term(tempdir())
}
}
\seealso{
\code{\link{fm}} to start a file manager.
}
