% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term}
\alias{term}
\title{Open a Terminal}
\usage{
term(path = getwd())
}
\arguments{
\item{path}{[\code{character(1)}]\cr
Path. Defaults to the current working directory as reported by \code{\link[base]{getwd}}.}
}
\description{
Opens a terminal emulator in directory \code{path}.

The possibilities to start a terminal are:
\describe{
  \item{\dQuote{tmux}:}{Create a tmux split if the R session runs inside tmux.}
  \item{\dQuote{rstudio}:}{Create a new RStudio terminal if the R session runs inside RStudio.}
  \item{\dQuote{opts}:}{If set, use the options \dQuote{gtfo.terminal.bin} and \dQuote{gtfo.terminal.args} to open a new terminal or terminal tab.}
  \item{\dQuote{default}:}{Open one of the hard-coded terminals. See note for details.}
}
The order of the different strategies can be controlled via the option \dQuote{gtfo.term.order} (default: \code{c("tmux", "rstudio", "opts", "default")}).
}
\note{
The default terminal will determined as follows:
\describe{
 \item{Mac OS:}{If called from \command{iTerm}, start \command{iTerm}. Start \command{Terminal} otherwise.}
 \item{Linux:}{Try to find the one of terminals in \code{$PATH} via \code{\link[base]{Sys.which}} and open the first one found:
   \command{kitty}, \command{alacritty}, \command{urxvt},
   \command{gnome-terminal}, \command{konsole},
   \command{xfce4-terminal}, \command{lxterminal}.}
 \item{Windows:}{Start powershell.}
}
}
\examples{
\dontrun{
# Terminal in R's working directory
term()

# Terminal in R's tempdir
term(tempdir())
}
}
\seealso{
\code{\link{fm}} to start a file manager and \code{\link{browse}} to start a browser.
}
